more exercises

+++ atividades/ex007.py
@@ -1,4 +1,4 @@
 nome = input('Digite seu nome:')
-n1 = int(input('Seja bem vindo {}, vamos calcular sua média, por favor digite sua primeira nota:'))
-n2 = int(input('Digite sua segunda nota:'))
+n1 = float(input('Seja bem vindo {}, vamos calcular sua média, por favor digite sua primeira nota:'))
+n2 = float(input('Digite sua segunda nota:'))
 print('Sua média é {}'.format((n1+n2)/2))
